feat!: only allow safe-precesion integers to pass `parsedInteger`
Fixes #743

This commit adds an additional check to the `parsedInteger` validator,
using `Number.isSafeInteger()` to require that any parsed integer has
maximum precision (cannot overflow). This better represents what the
average user would want, but may fail on some valid inputs. An
`unsafeParsedInteger` validator should be considered in the future.

Due to many upcoming changes to error scopes, 'string' was used as the
base type and the regex check was made in the narrow. The error message
when a non-string is passed has changed slightly, just saying "must be a
string" instead of "must be a well-formed integer string." This is a
very minor change and can be dismissed.

BREAKING CHANGE: Any unsafe integers that use the `parsedInteger`
validator will now fail the validation.

Signed-Off-By: LogN <logn@0xlogn.dev>