fix: accept newlines before an operator
Inserting newlines *after* an operator (`a | \n b`), currently
works, but placing the newline before it (`a \n | b`) fails, because
the dynamic operator parsing code is hardcoded to only skip spaces.
(The static parser *did* correctly check for any whitespace here.) In
addition, the list of terminating characters needs to include non-space
whitespace, otherwise the newlines can end up as part of the parsed
type.